% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testColumnsCosistency.R
\name{testColumnsCosistency}
\alias{testColumnsCosistency}
\title{Test data.frame based on some conditions}
\usage{
testColumnsCosistency(data, conditions)
}
\arguments{
\item{data}{data.frame}

\item{conditions}{character.vector "filter_ like" conditions, e.g. "Sepal.Length > 0"}
}
\value{
the same data.frame or raise error (what conditions were not met)
}
\examples{
testColumnsCosistency(iris, "Sepal.Length > 0")
iris$Sepal.Length[5:10] = NA
testColumnsCosistency(iris, "is.na(Sepal.Length)")
}
