% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAV_testers.R
\name{testIfNamesAreThere}
\alias{testIfNamesAreThere}
\title{Test column naming}
\usage{
testIfNamesAreThere(data, expected.columns)
}
\arguments{
\item{data}{data.frame you want to test}

\item{expected.columns}{character vector of expected column names}
}
\value{
same data.frame or raise error
}
\description{
Test column naming
}
