% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAV_testers.R
\name{testWhereAreNAs}
\alias{testWhereAreNAs}
\title{Test if you have NA's in the data.frame}
\usage{
testWhereAreNAs(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
the same data.frame or raise error
}
\description{
Test if you have NA's in the data.frame
}
