% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testWhereAreNAs.R
\name{testWhereAreNAs}
\alias{testWhereAreNAs}
\title{Test if you have NA's in the data.frame and if so - where}
\usage{
testWhereAreNAs(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
the same data.frame or raise error (exact positions of NAs)
}
\examples{
iris$Sepal.Length[5:10] = NA
testWhereAreNAs(iris)
}
